Refactor Start.js
